NetPlay: Supply every player's ping data to the client
This means that we now have feature parity with the combined server
before.

+++ Source/Core/Core/Src/NetPlayServer.h
@@ -61,7 +61,7 @@ private:
 		std::string		revision;
 
 		sf::SocketTCP	socket;
-		u64 ping;
+		u32 ping;
 		u32 current_game;
 	};
